mandatory part done, little refactoring and more testing needed, starting to work on bonus part. more tests added to pipex_tester

+++ pipex_tester/tester_loop.sh
@@ -5,11 +5,14 @@ nb_tests=${#cmd_original[@]}
 for (( i=0; i < $nb_tests; i++ )); do
 	expected_out=$(eval ${cmd_original[$i]} 2>&1)
 	test -e ${fileout[$i]} && expected_fileout=$(< ${fileout[$i]} cat) || expected_fileout=""
-	rm -f ${fileout[$i]}
-
+	if [ "${filein[$i]}" != "${fileout[$i]}" ]; then
+		rm -f ${fileout[$i]}
+	fi;
 	pipex_out=$(eval ${cmd_pipex[$i]} 2>&1)
 	test -e ${fileout[$i]} && pipex_fileout=$(< ${fileout[$i]} cat) || pipex_fileout=""
-	rm -f ${fileout[$i]}
+	if [ "${filein[$i]}" != "${fileout[$i]}" ]; then
+		rm -f ${fileout[$i]}
+	fi;
 	
 	TEST_CLR="$GRN"
 	OK_OUT=1
@@ -25,7 +28,8 @@ for (( i=0; i < $nb_tests; i++ )); do
 	fi;
 	
 	echo -e "\n-----------------------------------------"
-	echo -e "$TEST_CLR Test$i:$WHT ${cmd_original[$i]}"
+	echo -e "$TEST_CLR Test    $i:$WHT ${cmd_original[$i]}"
+	echo -e "$TEST_CLR Pipex cmd:$WHT ${cmd_pipex[$i]}"
 	if [ $OK_OUT == 0 ]; then
 		echo -e "$YEL Differences in program output:$WHT"
 		echo -e "$BLU bash output :$WHT"
@@ -34,7 +38,7 @@ for (( i=0; i < $nb_tests; i++ )); do
 		echo "$pipex_out"
 	fi;
 	if [ $OK_FILEOUT == 0 ]; then
-		echo -e "$YEL Differences in final file:$WHT"
+		echo -e "$RED Differences in final file:$WHT"
 		echo -e "${BLU}bash file :${WHT}"
 		echo "$expected_fileout"
 		echo -e "${BLU}pipex file:${WHT}"
